added role functions

+++ models/appDB.sql
@@ -12,7 +12,7 @@ CREATE TABLE department(
 CREATE TABLE roles(
 	id INT AUTO_INCREMENT NOT NULL,
     title VARCHAR(50) NOT NULL,
-    salary DECIMAL NOT NULL,
+    salary INT NOT NULL,
     department_id INT NOT NULL,
     PRIMARY KEY(id),
     FOREIGN KEY (department_id) REFERENCES department (id)
